ui: split offroad home in its own class
