+++ selfdrive/ui/qt/home.h
@@ -7,16 +7,15 @@
 #include <QTimer>
 #include <QWidget>
 
-#include "common/params.h"
-#include "selfdrive/ui/qt/body.h"
-#include "selfdrive/ui/qt/widgets/offroad_alerts.h"
 #include "selfdrive/ui/ui.h"
 
 #ifdef SUNNYPILOT
 #include "selfdrive/ui/sunnypilot/qt/widgets/controls.h"
 #include "selfdrive/ui/sunnypilot/qt/onroad/onroad_home.h"
+#include "selfdrive/ui/sunnypilot/qt/offroad/offroad_home.h"
 #include "selfdrive/ui/sunnypilot/qt/sidebar.h"
 #define OnroadWindow OnroadWindowSP
+#define OffroadHome OffroadHomeSP
 #define LayoutWidget LayoutWidgetSP
 #define Sidebar SidebarSP
 #define ElidedLabel ElidedLabelSP
@@ -26,30 +25,7 @@
 #include "selfdrive/ui/qt/sidebar.h"
 #endif
 
-class OffroadHome : public QFrame {
-  Q_OBJECT
-
-public:
-  explicit OffroadHome(QWidget* parent = 0);
-
-signals:
-  void openSettings(int index = 0, const QString &param = "");
-
-private:
-  void showEvent(QShowEvent *event) override;
-  void hideEvent(QHideEvent *event) override;
-  void refresh();
-
-  Params params;
-
-  QTimer* timer;
-  ElidedLabel* version;
-  QStackedLayout* center_layout;
-  UpdateAlert *update_widget;
-  OffroadAlert* alerts_widget;
-  QPushButton* alert_notif;
-  QPushButton* update_notif;
-};
+#include "selfdrive/ui/qt/offroad/offroad_home.h"
 
 class HomeWindow : public QWidget {
   Q_OBJECT
