+++ selfdrive/ui/sunnypilot/qt/offroad/offroad_home.cc
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) 2021-, Haibin Wen, sunnypilot, and a number of other contributors.
+ *
+ * This file is part of sunnypilot and is licensed under the MIT License.
+ * See the LICENSE.md file in the root directory for more details.
+ */
+
+#include "selfdrive/ui/sunnypilot/qt/offroad/offroad_home.h"
+
+OffroadHomeSP::OffroadHomeSP(QWidget *parent) : OffroadHome(parent) {
+}
